// src/app/dggma/interfaces/pi.interface.ts
export interface Pi {
    interview__key:   string;
    interview__id:    string;
    secuencia_pi__id: number;
    dg_pi:            number;
    nom_pi:           number;
    pp_pi:            number;
    dga_pi:           number;
}
export interface ProgInformacion {
    id:             number;
    text:           string;
    parentid:       number;
    attachmentname: string;
}
export interface ProcProduccion {
    id:             number;
    text:           string;
    parentid:       number;
    attachmentname: string;
}
export interface DgaPprod {
    id:             number;
    text:           string;
    parentid:       number;
    attachmentname: string;
}