// src/app/dggma/interfaces/aeg.interface.ts
export interface SecuenciaAeg {
    interview__key:    string;
    interview__id:     string;
    secuencia_aeg__id: number;
    dg_aeg:            number;
    nom_aeg:           number;
    dga_aeg:           number;
}
export interface Aeg2 {
    id:             number;
    text:           string;
    parentid:       number;
    attachmentname: string;
}
export interface DgaProd {
    id:             number;
    text:           string;
    parentid:       number;
    attachmentname: string;
}